Now parsing ESRI/GDAL NODATA tag as a direct double. Also removed a unit test for image conversion from tiles to rows which had an invalid number of bits per pixel. This case is covered in the GeoTiffReaderSpec instead so no functionality is less tested than before.

+++ raster/src/main/scala/geotrellis/raster/io/geotiff/reader/TagReader.scala
@@ -151,7 +151,8 @@ case class TagReader(byteBuffer: ByteBuffer) {
       case CopyrightTag => directory |-> copyrightLens set(Some(string))
       case AsciisTag => directory |-> asciisLens set(Some(string))
       case GDALInternalNoDataTag => directory |-> gdalInternalNoDataLens set(
-        Some(string.filter(_ != '\0').map(x => if (x.toByte == 3) '0' else x))
+        Some(string.filter(_ != '\0').map(x =>
+          if (x.toByte == 3) '0' else x).toDouble)
       )
       case tag => directory |-> asciisMapLens modify(_ + (tag -> string))
     }
